Add stream writer.
